+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/source/src/app/components/Base/Header/Header.js
@@ -18,7 +18,7 @@
 
 import React from 'react'
 import {Link} from "react-router-dom";
-import AuthManager from '../../../data/Auth.js';
+import AuthManager from '../../../data/AuthManager.js';
 
 const Header = (props) => {
     return (
