+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/source/test/auth/login.test.js
@@ -16,18 +16,18 @@
  * under the License.
  */
 
-import Auth from '../../src/app/data/Auth.js'
+import AuthManager from '../../src/app/data/AuthManager.js'
 import {describe, it} from "mocha";
 import {assert} from 'chai';
 import https from 'https';
 
-describe('Auth',
+describe('AuthManager',
     function () {
         describe('#authenticateUser()',
             function () {
                 it('Should return HTTP 200 status code if user authenticate',
                     function () {
-                        let authenticator = new Auth();
+                        let authenticator = new AuthManager();
                         const agent = new https.Agent({rejectUnauthorized: false});
                         let promised_auth = authenticator.authenticateUser('admin', 'admin');
                         return promised_auth.then((response) => {
