+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/source/test/utils.js
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-import Auth from "../src/app/data/Auth.js";
+import AuthManager from "../src/app/data/AuthManager.js";
 class TestUtils {
     static setupMockEnviroment() {
         global.window = {
@@ -44,7 +44,7 @@ class TestUtils {
     }
 
     static userLogin(username = 'admin', password = 'admin') {
-        let authenticator = new Auth();
+        let authenticator = new AuthManager();
         return authenticator.authenticateUser(username, password);
     }
 }
